refactor(runtimed): centralize path helpers in paths.rs
Path helpers were scattered across daemon.rs (5.7k lines) and
notebook_sync_server.rs (15.6k lines) with no shared home. Each caller
reached into a different upstream crate - runtimed-client for cache/pool
paths, runt-workspace for the socket, notebook-doc for the doc filename -
with direct absolute imports plus two local path helpers buried deep in
the god files. Small drift risk and a real navigation hazard.

Collect the single source of truth into crates/runtimed/src/paths.rs:

- Re-export default_socket_path, default_cache_dir, pool_env_root, and
  notebook_doc_filename so crate::paths::* is a single entry point for
  runtimed's own code. Upstream crates are untouched.
- Move snapshot_before_delete out of notebook_sync_server.rs.
- Move normalize_save_target out of notebook_sync_server.rs.
- Move looks_like_untitled_notebook_path out of daemon.rs and fold its
  inline #[cfg(test)] block into paths::tests.

Visibility preserved (private -> pub(crate) only where crossing the file
boundary). All callsites updated to go through crate::paths::*. No
logic changes. cargo test -p runtimed --lib (380 passed) and
cargo test -p runtimed --tests all green.
