+++ crates/runtimed/src/lib.rs
@@ -35,6 +35,7 @@ pub mod markdown_assets;
 pub mod notebook_sync_server;
 pub mod output_prep;
 pub mod output_store;
+pub mod paths;
 pub mod process_groups;
 pub mod project_file;
 pub mod runtime_agent;
